Initial commit
Includes a cli, a picture builder which determines the timestamps of
pictures, from EXIF and mtime, and an organizer which does nothing but
calculate a new path for each picture.

A bit larger than it should have been. But I've been fiddling around
a lot since I don't know the language, which is also the reason for the
lack of tests.